+++ .lintstagedrc.js
@@ -1,6 +1,6 @@
 module.exports = {
   "**/*.{ts,tsx}": [
     (filenames) => filenames.map((filename) => `prettier '${filename}' --write`),
-    (filenames) => (filenames.length > 1 ? "yarn lint" : `eslint ${filenames.join(" ")}`),
+    (filenames) => (filenames.length > 10 ? "yarn lint" : `eslint ${filenames.join(" ")}`),
   ],
 };