+++ src/components/Layout.tsx
@@ -1,14 +1,13 @@
 import React, { FC } from "react";
 import { Box, Flex } from "rebass/styled-components";
 
+import SideBar from "./SideBar";
+
 const Layout: FC = ({ children }) => (
-  <Box sx={{ maxWidth: 2048, mx: "auto", px: 3 }}>
+  <Box sx={{ maxWidth: 2048, mx: "auto" }} fontSize={[1, 2, 3]} pt={[3, 4, 5]}>
     <Flex flexDirection="row" flex-wrap="wrap">
-      <Box px={2} py={2} width={1 / 3} sx={{ borderRight: "1px solid ", borderColor: "lightgray" }}>
-        <Box mx="auto" />
-        {/* TODO: add SideBar */}
-      </Box>
-      <Box px={2} py={2} width={2 / 3}>
+      <SideBar />
+      <Box as="main" width={2 / 3}>
         {children}
       </Box>
     </Flex>