+++ src/constants/nav.ts
@@ -1,0 +1,19 @@
+interface NavOptions {
+  href: string;
+  title: string;
+}
+
+export const NAV_OPTIONS: NavOptions[] = [
+  {
+    href: "/about",
+    title: "About",
+  },
+  {
+    href: "/",
+    title: "Blog",
+  },
+  {
+    href: "/contact",
+    title: "Contact",
+  },
+];